v2/requirements.txt
# Options Breakout Tracker V2 - Dependencies

# Upstox API SDK (includes WebSocket support)
upstox-python-sdk>=2.0.0

# HTTP requests
requests>=2.28.0

# Async WebSocket server for dashboard
aiohttp>=3.8.0
aiohttp-cors>=0.7.0

# Configuration
pyyaml>=6.0

# Data handling
pandas>=1.5.0

# Protobuf (for WebSocket message parsing)
protobuf>=4.0.0
